changed setup script

+++ setup.sh
@@ -1,14 +1,15 @@
 #!/bin/bash
 
 user="$USER"
+arch=1
 
 if command -v pacman >/dev/null 2>&1 && pacman -Qq >/dev/null 2>&1; then
+  arch=0
   sudo pacman -S neovim kitty hyprland tmux unzip npm nodejs zoxide swww python-pywal zsh fzf ripgrep fd bat clang --noconfirm
 else 
   echo "You are not on an arch based distro. Only continue after installing the following packages: neovim, kitty, hyprland, tmux, unzip, npm, nodejs, zoxide, swww, python-pywal, zsh, fzf, ripgrep, fd, bat, clang before running"
 
-  echo -p "Contiune? (y/n) "
-  read continue
+  read -p "Contiune? (y/n) " continue
   if [[ $continue != "y" ]]; then
     exit 1
   fi
@@ -49,14 +50,16 @@ function update_dotfiles() {
   echo "Creating symlink: ${HOME}/.config/tmux"
   ln -sf "${HOME}/dotfiles/tmux" "${HOME}/.config/tmux"
 
-  # Manually check and remove system-wide symlink
-  if [[ -L /usr/bin/sys-maintenance ]]; then
-    echo "Removing symlink: /usr/bin/sys-maintenance"
-    sudo rm /usr/bin/sys-maintenance
-  fi
+  if [[ "$arch" -eq 0 ]]; then
+    # Manually check and remove system-wide symlink
+    if [[ -L /usr/bin/sys-maintenance ]]; then
+      echo "Removing symlink: /usr/bin/sys-maintenance"
+      sudo rm /usr/bin/sys-maintenance
+    fi
 
-  echo "Creating system-wide symlink: /usr/bin/sys-maintenance"
-  sudo ln -sf "${HOME}/dotfiles/sys-maintenance" /usr/bin/sys-maintenance
+    echo "Creating system-wide symlink: /usr/bin/sys-maintenance"
+    sudo ln -sf "${HOME}/dotfiles/sys-maintenance" /usr/bin/sys-maintenance
+  fi
 }
 
 update_dotfiles
